+++ ReleaseNotes/release_notes.md
@@ -1,3 +1,10 @@
+## New features:
+
+## Backward incompatible  
+
+## Bug fixes:
+
+
 # v1.6.9.2
 
 ## New features: